+++ api/src/main/java/org/allaymc/api/world/WorldState.java
@@ -1,0 +1,11 @@
+package org.allaymc.api.world;
+
+/**
+ * @author daoge_cmd
+ */
+public enum WorldState {
+    STARTING,
+    RUNNING,
+    STOPPING,
+    STOPPED
+}